Fix ISO single kernel update

+++ 4-update-iso-chroot.sh
@@ -2,16 +2,13 @@
 
 # Linuxium's scripts to create a custom Ubuntu ISO
 
-source include-chroot-variables.txt
+source chroot-variables.txt
 
-[ ! -f ${PATH_TO}/sources.list ] && echo "Using sources.list from git." && cp sources.list ${PATH_TO}/
+# include definition of additional packages to be installed
+sudo cp iso_packages.source iso-chroot/usr/src
 
-# copy list of desired packages
-sudo cp target-packages.sh iso-chroot/
-
-# add in kernel debs(remove debug ones)
-sudo bash -c "cp development-chroot/usr/src/*.deb iso-chroot/usr/src/"
-sudo bash -c "rm iso-chroot/usr/src/*-dbg_*.deb"
+# add in kernel debs
+sudo bash -c "cp deb_binaries/*.deb iso-chroot/usr/src/"
 
 # add in UCM files for sound
 sudo mkdir -p iso-chroot/usr/share/alsa/ucm
@@ -20,8 +17,8 @@ sudo cp -rf ${PATH_TO}/UCM-master/* iso-chroot/usr/share/alsa/ucm
 # install kernel debs and additional packages
 sudo mv iso-chroot/etc/apt/sources.list iso-chroot/etc/apt/sources.list.orig
 sudo cp ${PATH_TO}/sources.list iso-chroot/etc/apt/
-sudo rm iso-chroot/etc/resolv.conf
 sudo cp /etc/resolv.conf iso-chroot/etc/
+sudo mount --bind /dev/ iso-chroot/dev
 sudo chroot iso-chroot <<+
 mount -t proc none /proc
 mount -t sysfs none /sys
@@ -36,7 +33,8 @@ dpkg -i /usr/src/linux*.deb
 
 apt-get update && apt-get -y upgrade
 
-source /target-packages.sh
+source /usr/src/iso_packages.source
+rm /usr/src/iso_packages.source
 
 apt-get clean
 apt-get autoclean
@@ -65,23 +63,23 @@ sudo rm -rf iso-chroot/run
 sudo mkdir iso-chroot/dev
 sudo mkdir iso-chroot/run
 
-# update kernel in iso
-sudo cp iso-chroot/boot/vmlinuz-${LINUXIUM_KERNEL_RELEASE}-${LINUXIUM_KERNEL_VERSION} iso-directory-structure/casper/vmlinuz.efi
-sudo cp iso-chroot/boot/initrd.img-${LINUXIUM_KERNEL_RELEASE}-${LINUXIUM_KERNEL_VERSION} iso-directory-structure/casper/initrd.lz
-
-# unpack initramfs
-[ -d initrd ] && rm -rf initrd
-mkdir initrd
-cd initrd
-sudo dd status=none if=../iso-directory-structure/casper/initrd.lz | gzip -d | sudo cpio -id
-
-# add in UCM files for sound in initramfs
-sudo mkdir -p usr/share/alsa/ucm
-sudo cp -rf ${PATH_TO}/UCM-master/* usr/share/alsa/ucm
-
-# repack initramfs
-sudo rm -f ../iso-directory-structure/casper/initrd.lz
-sudo find | sudo cpio -o -H newc | gzip | sudo tee ../iso-directory-structure/casper/initrd.lz > /dev/null
-cd ..
-sudo rm -rf initrd
+# update the ISO
+if [ ${LINUXIUM_KERNEL_VERSION} != "linuxium" ]; then
+	# update kernel
+	sudo bash -c "cp iso-chroot/boot/vmlinuz-*-${LINUXIUM_KERNEL_VERSION:0:4}* iso-directory-structure/casper/vmlinuz.efi"
+	sudo bash -c "cp iso-chroot/boot/initrd.img-*-${LINUXIUM_KERNEL_VERSION:0:4}* iso-directory-structure/casper/initrd.lz"
+	# unpack initramfs
+	[ -d initrd ] && rm -rf initrd
+	mkdir initrd
+	cd initrd
+	sudo dd status=none if=../iso-directory-structure/casper/initrd.lz | gzip -d | sudo cpio -id
+	# add in UCM files for sound in initramfs
+	sudo mkdir -p usr/share/alsa/ucm
+	sudo cp -rf ${PATH_TO}/UCM-master/* usr/share/alsa/ucm
+	# repack initramfs
+	sudo rm -f ../iso-directory-structure/casper/initrd.lz
+	sudo find | sudo cpio -o -H newc | gzip | sudo tee ../iso-directory-structure/casper/initrd.lz > /dev/null
+	cd ..
+	sudo rm -rf initrd
+fi
